aihero/aihero/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "A hybrid search and QA agent over documentation with text + vector search"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Neidy Tunzine", email = "ntunzine@gmail.com" }
]
requires-python = ">=3.10,<3.13"
dependencies = [
  "minsearch>=0.0.5",
  "openai>=1.108.2",
  "pydantic-ai==1.0.9",
  "python-frontmatter>=1.1.0",
  "requests>=2.32.5",
  "numpy>=1.26.0",
  "tqdm>=4.66.0",
  "sentence-transformers>=3.2.0",
  "rich>=13.7.0",
  "streamlit>=1.50.0",
]


[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "black>=24.4.2",
  "flake8>=7.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
